Marketing Campaign
The data used for the coursework is marketing campaign dataset based on a case of a retailer company. The primary objective of your work is to prepare data for further data mining and analysis. The data set contains 1500 customer records. Each record consists of 19 variables, which includes socio-demographic and product ownership information.